experimental/requirements.txt
mlx-vlm
pillow
docling-core
